Remove circular dependency
sensors.py <-> sitelogtogeodesyml.py
